Move items between groups
This still doesn't cover all edge cases, but it's the first prototype
implementation for moving items between groups.

The problem main problem comes from firing to independent events at the same
time. Each event has only the view of it's own current state, so the later
propagate event will override the result of the first event.

To overcome this, we must provide a mechanism for serializing those to events
and ensure both child components get a change of receiving a consistent view
after the first event has updated the state.

Further, we'd like to merge those to mutations into 1 update event, so the
consume of the QueryBuilder will be only triggered once.

With the current implementation, we can move items between siblings.
However, moving items within the current hierarchy is still not working
properly...